Fix loading of plugin, add ping REST endpoint, run integration tests
Due to an incorrect Spring import instruction, the plugin could not be loaded in
Jira. This has now be fixed.

Furthermore, the same ping REST endpoint for checking the plugin availability
has been added, as in the bitbucket-confapi-plugin. It makes sense to add such a
resource and not to use the UPM for this purpose, as Crowd does not have the UPM
and so, the same approach for checking the plugin availability can be used
amongst all ConfAPI plugins.

And last, a simple integration test has been added using the new ping REST
endpoint in order to make sure automatically that the plugin can be loaded in
Jira.